support for nifti file format, extra parse args
use NiftiReader to read nifti files
added argparse parameters for rotations

+++ src/main.cpp
@@ -6,48 +6,41 @@ using namespace std;
 using namespace argparse;
 
 int main(int argc, const char **argv) {
-  DRRgenerator drrgene;
-
-  // we only load mhd file
   ArgumentParser parser("Parse CT file path");
-
-  parser.add_argument()
-      .names({"-f", "--file"})
-      .description("file path")
-      .required(true);
-
-  parser.add_argument()
-      .names({"-t", "--type"})
-      .description("file type")
-      .required(false);
   parser.enable_help();
+  parser.add_argument("-f", "--file", "file path", true);
+  parser.add_argument("-t", "--type", "file type");
+  parser.add_argument("-r", "--rot", "rotation x y z ");
+  parser.add_argument("-o", "--output", "output file name");
+  parser.add_argument("-s", "--size", "size of the output image");
+  parser.add_argument("-i", "--inverse",
+                      "invert x-ray? yes-> dark values for dense structure");
 
   auto err = parser.parse(argc, argv);
 
-  if (parser.exists("help")) {
-    parser.print_help();
-    return 0;
-  }
-  std::string CT_folder = "data/";
-  std::string raw_filename = CT_folder + "FullHead.raw";
-  std::string info_filename = CT_folder + "FullHead.mhd";
+  bool invert = false;
+  if (parser.exists("i")) invert = true;
+  string filename = parser.get<std::string>("f");
 
-  const std::string DEFAULT_FILETYPE = "MHD";
-  std::string filetype = DEFAULT_FILETYPE;
-
-  if (parser.exists("t")) {
-    filetype = parser.get<std::string>("t");
+  float roll = 0.0, pitch = 0.0, yaw = 0.0;
+  if (parser.exists("r")) {
+    auto rotations = parser.get<std::vector<float>>("r");
+    roll = rotations[0];
+    pitch = rotations[1];
+    yaw = rotations[2];
   }
 
-  if (parser.exists("f")) {
-    info_filename = parser.get<std::string>("f");
-    info_filename = info_filename.substr(0, info_filename.rfind(".")) + ".mhd";
-    raw_filename = info_filename.substr(0, info_filename.rfind(".")) + ".raw";
+  int sx = 512, sy = 512;
+  if (parser.exists("s")) {
+    std::vector<int> size = parser.get<std::vector<int>>("s");
+    sx = size[0];
+    sy = size[1];
+    std::for_each(size.begin(), size.end(),
+                  [](const int &e) { std::cout << e << " "; });
   }
-
   const auto startct = chrono::system_clock::now();
-
-  drrgene.load_CT(raw_filename, info_filename);
+  DRRgenerator drrgene(roll, pitch, yaw, sx, sy, invert);
+  drrgene.load_CT(filename);
 
   const auto stopct = chrono::system_clock::now();
   const auto duratct =
@@ -61,15 +54,21 @@ int main(int argc, const char **argv) {
   const auto durat =
       chrono::duration_cast<chrono::milliseconds>(stop - start).count();
   std::cout << "DRR created in " << durat << " ms" << std::endl;
+
   cv::namedWindow("DRR", CV_WINDOW_KEEPRATIO);
   cv::imshow("DRR", color);
 
-  std::string out_filename =
-      info_filename.substr(0, info_filename.rfind(".")) + ".png";
-  std::cout << "Writing DRR to file " << out_filename << std::endl;
+  if (parser.exists("output")) {
+    std::string out_filename = "output.png";
+    if (parser.exists("o")) {
+      out_filename = parser.get<std::string>("o");
+    }
+    std::cout << "Writing DRR to file " << out_filename << std::endl;
 
-  std::cout << "Matrix size:" << color.size() << std::endl;
-  cv::imwrite(out_filename, color);
+    std::cout << "Matrix size:" << color.size() << std::endl;
+    cv::imwrite(out_filename, color);
+  }
+
   cv::waitKey(0);
   return 0;
 }
